Fix column order
Django tables have some weird rules around field order, which is why one
might specify a sequence explicitly.  Unfortunately the way sequence was
being defined for the summary submissions table (the one shown on the
dashboard) caused the order to fall back to wahtever the default field
order was.

That order is overrode by the submissions table (the one shown in
submissions).

In order to have the two tables match order we need to define the
sequences in the same way.

Issue #29