Remove filter and search from recent tables
The recent tables were designed to only show the five most recent
reviews / submissions, yet they were presented to the user as though
they were showing "all".

This was confusing because the full table functionality was limited --
in particular, sort wasn't possible due to the fact that it was
a subset.  There were UX components for search and filter but those were
misleading because they actually took the user to another page.

By removing the interactive components we make the table purpose more
explicit.  The user can navigate to a more full featured table if there
is enough data to warrant.

Issue #30