Remove determination indentation
Unfortunately jinja doesn't think of whitespace the same way a code file
would; in particular, indentation in output is considered part of the
presented output.

This isn't a problem in HTML, but for plaintext emails it means our
whitespace makes it into the final email / the end email client.
Stripping out all indendation matches the indended rendering.

Furthermore I added whitepsace stripping to all of the logical blocks
(`{%- -%}` instead of `{% %}`), which should prevent that whitespace
from showing up.

Issue #61