Fix change email error
Users without a `slack` field could not change their email address due
to the fact that doing so would attempt to set the slack column to null
(and nulls are not allowed).

This updates the save of the change email field to only modify the slack
value if it exists.

Issue #2784 Changing email on non-staff accounts creates non-null
            constraint violation