Remove unwanted submissions fields
We don't want the fund, round, or lead information to render on the
various submissions tables.  It looks like the way that's done in Django
is to actually delete the populated attributes (if you just take it out
of the field list the fields are appended at the end).

Issue #94