Add new recommendation values
The recommendation field is required and previously only allowed for
three options ("Yes" "No" and "Maybe").  These three values were used to
calculate a "stop light" icon that would be red, green, or yellow
depending on whether there was consensus in a given direction for the
application.

ARDC wants more granularity in the recommendation field.  There were
a few options discussed, including the creation of a new field that
would allow recommendation to stay, essentially, "yes no maybe" and
collect things like "strong" and "weak" weights separately.

Sticking to one field is likely a better user experience.  As such we're
introducing the concept of a mapping from field values in to Y/N/M
values.

The field itself isn't configurable by an end user, and I don't think
that functionality is necessarily important.  It would be good to
eventually move this kind of functionality out of the code and into
settings, however.

Issue #96 Make review recommendation field configurable