Fix error where limit_statuses was duplicated
Related to commit 7467aa8c6, where the limit statuses needs to be set so
that we can reduce the statuses used.  When that specified list was
being passed in with a list already in the kwargs, it then got specified
twice in the parent's constructor.  This happened on the other
submission pages (received, reviewed, etc).